Remove alpnproxytest, copy relevant code to gatewaytest
This reverts commit d0bb77eb2.
