+++ lib/srv/alpnproxy/conn_test.go
@@ -26,8 +26,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/gravitational/teleport/lib/srv/alpnproxytest"
 )
 
 func TestPingConnection(t *testing.T) {
@@ -339,9 +337,7 @@ func makeTLSConn(t *testing.T, server, client net.Conn) (*tls.Conn, *tls.Conn) {
 	// Server
 	go func() {
 		tlsConn := tls.Server(server, &tls.Config{
-			Certificates: []tls.Certificate{
-				alpnproxytest.MustGenCertSignedWithCA(t, alpnproxytest.MustGenSelfSignedCert(t)),
-			},
+			Certificates: []tls.Certificate{mustGenCertSignedWithCA(t, mustGenSelfSignedCert(t))},
 		})
 		tlsConnChan <- struct {
 			*tls.Conn
