+++ lib/srv/alpnproxy/helpers_test.go
@@ -18,8 +18,12 @@ package alpnproxy
 
 import (
 	"context"
+	"crypto/rand"
+	"crypto/rsa"
 	"crypto/tls"
 	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
 	"fmt"
 	"io"
 	"net"
@@ -34,8 +38,8 @@ import (
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/auth"
+	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/srv/alpnproxy/common"
-	"github.com/gravitational/teleport/lib/srv/alpnproxytest"
 	"github.com/gravitational/teleport/lib/tlsca"
 )
 
@@ -49,7 +53,7 @@ type Suite struct {
 }
 
 func NewSuite(t *testing.T) *Suite {
-	ca := alpnproxytest.MustGenSelfSignedCert(t)
+	ca := mustGenSelfSignedCert(t)
 	pool := x509.NewCertPool()
 	pool.AddCert(ca.Cert)
 	l, err := net.Listen("tcp", "127.0.0.1:0")
@@ -104,7 +108,7 @@ func (s *Suite) GetCertPool() *x509.CertPool {
 }
 
 func (s *Suite) CreateProxyServer(t *testing.T) *Proxy {
-	serverCert := alpnproxytest.MustGenCertSignedWithCA(t, s.ca)
+	serverCert := mustGenCertSignedWithCA(t, s.ca)
 	tlsConfig := &tls.Config{
 		NextProtos: common.ProtocolsToString(common.SupportedProtocols),
 		ClientAuth: tls.VerifyClientCertIfGiven,
@@ -143,6 +147,68 @@ func (s *Suite) Start(t *testing.T) {
 		err := svr.Close()
 		require.NoError(t, err)
 	})
+}
+
+func mustGenSelfSignedCert(t *testing.T) *tlsca.CertAuthority {
+	caKey, caCert, err := tlsca.GenerateSelfSignedCA(pkix.Name{
+		CommonName: "localhost",
+	}, []string{"localhost"}, defaults.CATTL)
+	require.NoError(t, err)
+
+	ca, err := tlsca.FromKeys(caCert, caKey)
+	require.NoError(t, err)
+	return ca
+}
+
+type signOptions struct {
+	identity tlsca.Identity
+	clock    clockwork.Clock
+}
+
+func withIdentity(identity tlsca.Identity) signOptionsFunc {
+	return func(o *signOptions) {
+		o.identity = identity
+	}
+}
+
+func withClock(clock clockwork.Clock) signOptionsFunc {
+	return func(o *signOptions) {
+		o.clock = clock
+	}
+}
+
+type signOptionsFunc func(o *signOptions)
+
+func mustGenCertSignedWithCA(t *testing.T, ca *tlsca.CertAuthority, opts ...signOptionsFunc) tls.Certificate {
+	options := signOptions{
+		identity: tlsca.Identity{Username: "test-user"},
+		clock:    clockwork.NewRealClock(),
+	}
+
+	for _, opt := range opts {
+		opt(&options)
+	}
+
+	subj, err := options.identity.Subject()
+	require.NoError(t, err)
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+
+	tlsCert, err := ca.GenerateCertificate(tlsca.CertificateRequest{
+		Clock:     options.clock,
+		PublicKey: privateKey.Public(),
+		Subject:   subj,
+		NotAfter:  options.clock.Now().UTC().Add(time.Minute),
+		DNSNames:  []string{"localhost", "*.localhost"},
+	})
+	require.NoError(t, err)
+
+	keyRaw := x509.MarshalPKCS1PrivateKey(privateKey)
+	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: keyRaw})
+	cert, err := tls.X509KeyPair(tlsCert, keyPEM)
+	require.NoError(t, err)
+	return cert
 }
 
 func mustReadFromConnection(t *testing.T, conn net.Conn, want string) {
