+++ lib/srv/alpnproxy/local_proxy_test.go
@@ -42,7 +42,6 @@ import (
 
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/srv/alpnproxy/common"
-	"github.com/gravitational/teleport/lib/srv/alpnproxytest"
 	"github.com/gravitational/teleport/lib/tlsca"
 )
 
@@ -376,13 +375,13 @@ func TestCheckDBCerts(t *testing.T) {
 	for _, tt := range tests {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			tlsCert := alpnproxytest.MustGenCertSignedWithCA(t, suite.ca,
-				alpnproxytest.WithIdentity(tlsca.Identity{
+			tlsCert := mustGenCertSignedWithCA(t, suite.ca,
+				withIdentity(tlsca.Identity{
 					Username:        "test-user",
 					Groups:          []string{"test-group"},
 					RouteToDatabase: dbRouteInCert,
 				}),
-				alpnproxytest.WithClock(tt.clock),
+				withClock(tt.clock),
 			)
 			lp.SetCerts([]tls.Certificate{tlsCert})
 			tt.errAssertFn(t, lp.CheckDBCerts(tt.dbRoute))
