+++ lib/teleterm/gateway/gateway_test.go
@@ -25,7 +25,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/gravitational/teleport/lib/defaults"
-	"github.com/gravitational/teleport/lib/srv/alpnproxytest"
 	"github.com/gravitational/teleport/lib/teleterm/api/uri"
 	"github.com/gravitational/teleport/lib/teleterm/gatewaytest"
 	"github.com/gravitational/teleport/lib/tlsca"
@@ -54,7 +53,7 @@ func TestGatewayStart(t *testing.T) {
 		hs.Close()
 	})
 
-	keyPairPaths := gatewaytest.MustGenAndSaveCert(t, alpnproxytest.WithIdentity(tlsca.Identity{
+	keyPairPaths := gatewaytest.MustGenAndSaveCert(t, tlsca.Identity{
 		Username: "alice",
 		Groups:   []string{"test-group"},
 		RouteToDatabase: tlsca.RouteToDatabase{
@@ -62,7 +61,7 @@ func TestGatewayStart(t *testing.T) {
 			Protocol:    defaults.ProtocolPostgres,
 			Username:    "alice",
 		},
-	}))
+	})
 
 	gateway, err := New(
 		Config{
@@ -160,7 +159,7 @@ func createGateway(t *testing.T, tcpPortAllocator TCPPortAllocator) *Gateway {
 		hs.Close()
 	})
 
-	keyPairPaths := gatewaytest.MustGenAndSaveCert(t, alpnproxytest.WithIdentity(tlsca.Identity{
+	keyPairPaths := gatewaytest.MustGenAndSaveCert(t, tlsca.Identity{
 		Username: "alice",
 		Groups:   []string{"test-group"},
 		RouteToDatabase: tlsca.RouteToDatabase{
@@ -168,7 +167,7 @@ func createGateway(t *testing.T, tcpPortAllocator TCPPortAllocator) *Gateway {
 			Protocol:    defaults.ProtocolPostgres,
 			Username:    "alice",
 		},
-	}))
+	})
 
 	gateway, err := New(
 		Config{
