+++ lib/teleterm/gatewaytest/helpers.go
@@ -15,8 +15,11 @@
 package gatewaytest
 
 import (
+	"crypto/rand"
 	"crypto/rsa"
+	"crypto/tls"
 	"crypto/x509"
+	"crypto/x509/pkix"
 	"encoding/pem"
 	"fmt"
 	"net"
@@ -25,10 +28,12 @@ import (
 	"time"
 
 	"github.com/gravitational/trace"
+	"github.com/jonboulle/clockwork"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/exp/slices"
 
-	"github.com/gravitational/teleport/lib/srv/alpnproxytest"
+	"github.com/gravitational/teleport/lib/defaults"
+	"github.com/gravitational/teleport/lib/tlsca"
 )
 
 const timeout = time.Second * 5
@@ -140,14 +145,14 @@ type KeyPairPaths struct {
 	KeyPath  string
 }
 
-func MustGenAndSaveCert(t *testing.T, opts ...alpnproxytest.SignOptionsFunc) KeyPairPaths {
+func MustGenAndSaveCert(t *testing.T, identity tlsca.Identity) KeyPairPaths {
 	t.Helper()
 
 	dir := t.TempDir()
 
-	ca := alpnproxytest.MustGenSelfSignedCert(t)
+	ca := mustGenCACert(t)
 
-	tlsCert := alpnproxytest.MustGenCertSignedWithCA(t, ca, opts...)
+	tlsCert := mustGenCertSignedWithCA(t, ca, identity)
 
 	privateKey, ok := tlsCert.PrivateKey.(*rsa.PrivateKey)
 	if !ok {
@@ -180,4 +185,39 @@ func MustGenAndSaveCert(t *testing.T, opts ...alpnproxytest.SignOptionsFunc) Key
 		CertPath: certFile.Name(),
 		KeyPath:  keyFile.Name(),
 	}
+}
+
+func mustGenCACert(t *testing.T) *tlsca.CertAuthority {
+	caKey, caCert, err := tlsca.GenerateSelfSignedCA(pkix.Name{
+		CommonName: "localhost",
+	}, []string{"localhost"}, defaults.CATTL)
+	require.NoError(t, err)
+
+	ca, err := tlsca.FromKeys(caCert, caKey)
+	require.NoError(t, err)
+	return ca
+}
+
+func mustGenCertSignedWithCA(t *testing.T, ca *tlsca.CertAuthority, identity tlsca.Identity) tls.Certificate {
+	clock := clockwork.NewRealClock()
+	subj, err := identity.Subject()
+	require.NoError(t, err)
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+
+	tlsCert, err := ca.GenerateCertificate(tlsca.CertificateRequest{
+		Clock:     clock,
+		PublicKey: privateKey.Public(),
+		Subject:   subj,
+		NotAfter:  clock.Now().UTC().Add(time.Minute),
+		DNSNames:  []string{"localhost", "*.localhost"},
+	})
+	require.NoError(t, err)
+
+	keyRaw := x509.MarshalPKCS1PrivateKey(privateKey)
+	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: keyRaw})
+	cert, err := tls.X509KeyPair(tlsCert, keyPEM)
+	require.NoError(t, err)
+	return cert
 }
